requirements.txt
google-adk
requests>=2.31.0